--- a/pkg/controller/admin/email_test.go
+++ b/pkg/controller/admin/email_test.go
@@ -28,7 +28,9 @@ import (
 )
 
 func TestShowAdminEmail(t *testing.T) {
-	harness := envstest.NewServer(t)
+	t.Parallel()
+
+	harness := envstest.NewServer(t, testDatabaseInstance)
 
 	// Get the default realm
 	realm, err := harness.Database.FindRealm(1)