--- a/pkg/controller/apikey/create_test.go
+++ b/pkg/controller/apikey/create_test.go
@@ -29,7 +29,9 @@ import (
 )
 
 func TestHandleCreate(t *testing.T) {
-	harness := envstest.NewServer(t)
+	t.Parallel()
+
+	harness := envstest.NewServer(t, testDatabaseInstance)
 
 	// Get the default realm
 	realm, err := harness.Database.FindRealm(1)