--- a/pkg/database/database_util.go
+++ b/pkg/database/database_util.go
@@ -17,11 +17,15 @@ package database
 import (
 	"context"
 	"crypto/rand"
+	"encoding/hex"
+	"flag"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
 	"strconv"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
@@ -32,91 +36,151 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/jinzhu/gorm"
 	"github.com/ory/dockertest"
+	"github.com/ory/dockertest/docker"
 	"github.com/sethvargo/go-envconfig"
+
+	// imported to register the postgres migration driver
+	_ "github.com/golang-migrate/migrate/v4/database/postgres"
+	// imported to register the "file" source migration driver
+	_ "github.com/golang-migrate/migrate/v4/source/file"
+	// imported to register the "postgres" database driver for migrate
 )
 
 const (
+	// databaseName is the name of the template database to clone.
+	databaseName = "test-db-template"
+
+	// databaseUser and databasePassword are the username and password for
+	// connecting to the database. These values are only used for testing.
+	databaseUser     = "test-user"
+	databasePassword = "testing123"
+
+	// defaultPostgresImageRef is the default database container to use if none is
+	// specified.
 	defaultPostgresImageRef = "postgres:13-alpine"
 )
 
 var (
-	approxTime = cmp.Options{cmpopts.EquateApproxTime(time.Second)}
+	// ApproxTime is a compare helper for clock skew.
+	ApproxTime = cmp.Options{cmpopts.EquateApproxTime(1 * time.Second)}
 )
 
-// NewTestDatabaseWithCacher creates a database configured with a cacher for use
-// in testing.
+// TestInstance is a wrapper around the Docker-based database instance.
+type TestInstance struct {
+	pool      *dockertest.Pool
+	container *dockertest.Resource
+
+	db     *Database
+	dbLock sync.Mutex
+
+	skipReason string
+}
+
+// MustTestInstance is NewTestInstance, except it prints errors to stderr and
+// calls os.Exit when finished. Callers can call Close or MustClose().
+func MustTestInstance() *TestInstance {
+	testDatabaseInstance, err := NewTestInstance()
+	if err != nil {
+		fmt.Fprintln(os.Stderr, err.Error())
+		os.Exit(1)
+	}
+	return testDatabaseInstance
+}
+
+// NewTestInstance creates a new Docker-based database instance. It also creates
+// an initial database, runs the migrations, and sets that database as a
+// template to be cloned by future tests.
+//
+// This should not be used outside of testing, but it is exposed in the package
+// so it can be shared with other packages. It should be called and instantiated
+// in TestMain.
 //
 // All database tests can be skipped by running `go test -short` or by setting
 // the `SKIP_DATABASE_TESTS` environment variable.
-func NewTestDatabaseWithCacher(tb testing.TB, cacher cache.Cacher) (*Database, *Config) {
-	tb.Helper()
+func NewTestInstance() (*TestInstance, error) {
+	// Querying for -short requires flags to be parsed.
+	if !flag.Parsed() {
+		flag.Parse()
+	}
 
+	// Do not create an instance in -short mode.
 	if testing.Short() {
-		tb.Skipf("🚧 Skipping database tests (short)!")
+		return &TestInstance{
+			skipReason: "🚧 Skipping database tests (-short flag provided)!",
+		}, nil
 	}
 
+	// Do not create an instance if database tests are explicitly skipped.
 	if skip, _ := strconv.ParseBool(os.Getenv("SKIP_DATABASE_TESTS")); skip {
-		tb.Skipf("🚧 Skipping database tests (SKIP_DATABASE_TESTS is set)!")
+		return &TestInstance{
+			skipReason: "🚧 Skipping database tests (SKIP_DATABASE_TESTS is set)!",
+		}, nil
 	}
 
-	// Context.
 	ctx := context.Background()
 
-	// Create the pool (docker instance).
+	// Create the pool.
 	pool, err := dockertest.NewPool("")
 	if err != nil {
-		tb.Fatalf("failed to create Docker pool: %s", err)
+		return nil, fmt.Errorf("failed to create database docker pool: %w", err)
 	}
 
 	// Determine the container image to use.
-	repo, tag := postgresRepo(tb)
+	repository, tag, err := postgresRepo()
+	if err != nil {
+		return nil, fmt.Errorf("failed to determine database repository: %w", err)
+	}
 
-	// These credentials (and this entire file) are only used for tests. They are
-	// used when spinning up an in-memory database for tests.
-	dbname, username, password := "en-verification-server", "my-username", "abcd1234"
-
-	// Start the container.
+	// Start the actual container.
 	container, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: repo,
+		Repository: repository,
 		Tag:        tag,
 		Env: []string{
 			"LANG=C",
-			"POSTGRES_DB=" + dbname,
-			"POSTGRES_USER=" + username,
-			"POSTGRES_PASSWORD=" + password,
+			"POSTGRES_DB=" + databaseName,
+			"POSTGRES_USER=" + databaseUser,
+			"POSTGRES_PASSWORD=" + databasePassword,
 		},
+	}, func(c *docker.HostConfig) {
+		c.AutoRemove = true
+		c.RestartPolicy = docker.RestartPolicy{Name: "no"}
 	})
 	if err != nil {
-		tb.Fatalf("failed to start postgres container: %s", err)
+		return nil, fmt.Errorf("failed to start database container: %w", err)
 	}
 
-	// Force the database container to stop.
-	if err := container.Expire(30); err != nil {
-		tb.Fatalf("failed to force-stop container: %v", err)
+	// Stop the container after its been running for too long. No since test suite
+	// should take super long.
+	if err := container.Expire(120); err != nil {
+		return nil, fmt.Errorf("failed to expire database container: %w", err)
 	}
 
-	// Ensure container is cleaned up.
-	tb.Cleanup(func() {
-		if err := pool.Purge(container); err != nil {
-			tb.Fatalf("failed to cleanup postgres container: %s", err)
-		}
-	})
+	// Generate keys.
+	apiKeyDatabaseHMAC, err := generateKeys(2, 128)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate api key database hmac: %w", err)
+	}
+	apiKeySignatureHMAC, err := generateKeys(2, 128)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate api key signature hmac: %w", err)
+	}
+	verificationCodeDatabaseHMAC, err := generateKeys(2, 128)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate verification code database hmac: %w", err)
+	}
 
-	// Get the host. On Mac, Docker runs in a VM.
-	host := container.GetBoundIP("5432/tcp")
-	port := container.GetPort("5432/tcp")
-
-	// build database config.
+	// Build database configuration. This is required to connect to the database
+	// and to run the initial migrations.
 	config := &Config{
-		APIKeyDatabaseHMAC:           generateKeys(tb, 3, 128),
-		APIKeySignatureHMAC:          generateKeys(tb, 3, 128),
-		VerificationCodeDatabaseHMAC: generateKeys(tb, 3, 128),
+		APIKeyDatabaseHMAC:           apiKeyDatabaseHMAC,
+		APIKeySignatureHMAC:          apiKeySignatureHMAC,
+		VerificationCodeDatabaseHMAC: verificationCodeDatabaseHMAC,
 
-		User:     username,
-		Port:     port,
-		Host:     host,
-		Name:     dbname,
-		Password: password,
+		Host:     container.GetBoundIP("5432/tcp"),
+		Port:     container.GetPort("5432/tcp"),
+		User:     databaseUser,
+		Name:     databaseName,
+		Password: databasePassword,
 		SSLMode:  "disable",
 		Secrets: secrets.Config{
 			SecretManagerType: secrets.SecretManagerTypeInMemory,
@@ -127,96 +191,204 @@ func NewTestDatabaseWithCacher(tb testing.TB, cacher cache.Cacher) (*Database, *
 		},
 	}
 
-	// Wait for the container to start - we'll retry connections in a loop below,
-	// but there's no point in trying immediately.
-	time.Sleep(1 * time.Second)
-
-	// Load the configuration
+	// Parse configuration and override with test data.
 	db, err := config.Load(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to load database configuration: %w", err)
+	}
+
+	// Try to establish a connection to the database.
+	if err := db.Open(ctx); err != nil {
+		return nil, fmt.Errorf("failed to open database connection: %w", err)
+	}
+
+	// Run database migrations.
+	if err := runMigrations(db); err != nil {
+		return nil, fmt.Errorf("failed to run migrations: %w", err)
+	}
+
+	// Return the instance.
+	return &TestInstance{
+		pool:      pool,
+		container: container,
+		db:        db,
+	}, nil
+}
+
+// MustClose is like Close except it prints the error to stderr and calls os.Exit.
+func (i *TestInstance) MustClose() error {
+	if err := i.Close(); err != nil {
+		fmt.Fprintln(os.Stderr, err.Error())
+		os.Exit(1)
+	}
+	return nil
+}
+
+// Close terminates the test database instance, cleaning up any resources.
+func (i *TestInstance) Close() (retErr error) {
+	// Do not attempt to close  things when there's nothing to close.
+	if i.skipReason != "" {
+		return
+	}
+
+	defer func() {
+		if err := i.pool.Purge(i.container); err != nil {
+			retErr = fmt.Errorf("failed to purge database container: %w", err)
+			return
+		}
+	}()
+
+	if err := i.db.Close(); err != nil {
+		retErr = fmt.Errorf("failed to close connection: %w", err)
+		return
+	}
+
+	return
+}
+
+// NewDatabase creates a new database suitable for use in testing. It returns an
+// established database connection and the configuration.
+func (i *TestInstance) NewDatabase(tb testing.TB, cacher cache.Cacher) (*Database, *Config) {
+	tb.Helper()
+
+	// Ensure we should actually create the database.
+	if i.skipReason != "" {
+		tb.Skip(i.skipReason)
+	}
+
+	// Clone the template database.
+	newDatabaseName, err := i.clone()
 	if err != nil {
 		tb.Fatal(err)
 	}
 
+	// Build the new connection to the new database name.
+	config := i.db.config
+	config.Name = newDatabaseName
+
+	// Create shared context.
+	ctx := context.Background()
+
+	// Parse configuration and override with test data.
+	db, err := config.Load(ctx)
+	if err != nil {
+		tb.Fatalf("failed to load database configuration: %s", err)
+	}
 	db.keyManager = keys.TestKeyManager(tb)
 	db.config.EncryptionKey = keys.TestEncryptionKey(tb, db.keyManager)
 
+	// Try to establish a connection to the database.
 	if err := db.OpenWithCacher(ctx, cacher); err != nil {
-		tb.Fatal(err)
+		tb.Fatalf("failed to open database connection: %s", err)
 	}
 
-	// Disable logging temporarily for migrations. The callback registration is
-	// really quite chatty.
-	db.db.SetLogger(gorm.Logger{LogWriter: log.New(ioutil.Discard, "", 0)})
-	db.db = db.db.LogMode(false)
-
-	if err := db.RunMigrations(ctx); err != nil {
-		tb.Fatalf("failed to migrate database: %v", err)
-	}
-
-	// Re-enable logging.
-	db.db.SetLogger(gorm.Logger{LogWriter: log.New(os.Stdout, "", 0)})
-
-	// Close db when done.
+	// Close connection and delete database when done.
 	tb.Cleanup(func() {
-		if err := db.db.Close(); err != nil {
-			tb.Fatal(err)
+		// Close connection first. It is an error to drop a database with active
+		// connections.
+		if err := db.Close(); err != nil {
+			tb.Errorf("failed to close database connection: %s", err)
+		}
+
+		// Drop the database to keep the container from running out of resources.
+		q := fmt.Sprintf(`DROP DATABASE IF EXISTS "%s" WITH (FORCE);`, newDatabaseName)
+
+		i.dbLock.Lock()
+		defer i.dbLock.Unlock()
+
+		if err := i.db.db.Exec(q).Error; err != nil {
+			tb.Errorf("failed to drop database %q: %s", newDatabaseName, err)
 		}
 	})
 
 	return db, config
 }
 
-// NewTestDatabaseWithConfig creates a new database suitable for use in testing.
-// This should not be used outside of testing, but it is exposed in the main
-// package so it can be shared with other packages.
-//
-// All database tests can be skipped by running `go test -short` or by setting
-// the `SKIP_DATABASE_TESTS` environment variable.
-func NewTestDatabaseWithConfig(tb testing.TB) (*Database, *Config) {
-	tb.Helper()
+// clone creates a new database with a random name from the template instance.
+func (i *TestInstance) clone() (string, error) {
+	// Generate a random database name.
+	name, err := randomDatabaseName()
+	if err != nil {
+		return "", fmt.Errorf("failed to generate random database name: %w", err)
+	}
 
-	return NewTestDatabaseWithCacher(tb, nil)
+	// Setup context and create SQL command. Unfortunately we cannot use parameter
+	// injection here as that's only valid for prepared statements, for which this
+	// is not. Fortunately both inputs can be trusted in this case.
+	q := fmt.Sprintf(`CREATE DATABASE "%s" WITH TEMPLATE "%s";`, name, databaseName)
+
+	// Unfortunately postgres does not allow parallel database creation from the
+	// same template, so this is guarded with a lock.
+	i.dbLock.Lock()
+	defer i.dbLock.Unlock()
+
+	// Clone the template database as the new random database name.
+	if err := i.db.db.Exec(q).Error; err != nil {
+		return "", fmt.Errorf("failed to clone template database: %w", err)
+	}
+	return name, nil
 }
 
-// NewTestDatabase creates a new test database with the defautl configuration.
-//
-// All database tests can be skipped by running `go test -short` or by setting
-// the `SKIP_DATABASE_TESTS` environment variable.
-func NewTestDatabase(tb testing.TB) *Database {
-	tb.Helper()
+// runMigrations runs the migrations for the database.
+func runMigrations(db *Database) error {
+	ctx := context.Background()
 
-	db, _ := NewTestDatabaseWithConfig(tb)
-	return db
+	// Disable logging for migrations - the callback registration is quite chatty.
+	db.db.SetLogger(gorm.Logger{LogWriter: log.New(ioutil.Discard, "", 0)})
+	db.db = db.db.LogMode(false)
+
+	defer func() {
+		// Re-enable logging.
+		db.db.SetLogger(gorm.Logger{LogWriter: log.New(os.Stdout, "", 0)})
+	}()
+
+	// Run migrations.
+	if err := db.RunMigrations(ctx); err != nil {
+		return fmt.Errorf("failed to migrate database: %v", err)
+	}
+
+	return nil
 }
 
-func generateKeys(tb testing.TB, qty, length int) []envconfig.Base64Bytes {
-	tb.Helper()
+// postgresRepo returns the postgres container image name based on an
+// environment variable, or the default value if the environment variable is
+// unset.
+func postgresRepo() (string, string, error) {
+	ref := os.Getenv("CI_POSTGRES_IMAGE")
+	if ref == "" {
+		ref = defaultPostgresImageRef
+	}
 
+	parts := strings.SplitN(ref, ":", 2)
+	if len(parts) != 2 {
+		return "", "", fmt.Errorf("invalid reference for database container: %q", ref)
+	}
+	return parts[0], parts[1], nil
+}
+
+// randomDatabaseName returns a random database name.
+func randomDatabaseName() (string, error) {
+	b := make([]byte, 4)
+	if _, err := rand.Read(b); err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(b), nil
+}
+
+// generateKeys creates qty keys of the specified length.
+func generateKeys(qty, length int) ([]envconfig.Base64Bytes, error) {
 	keys := make([]envconfig.Base64Bytes, 0, qty)
 	for i := 0; i < qty; i++ {
 		buf := make([]byte, length)
 		n, err := rand.Read(buf)
 		if err != nil {
-			tb.Fatal(err)
+			return nil, fmt.Errorf("failed to read random: %w", err)
 		}
 		if n < length {
-			tb.Fatalf("insufficient bytes read: %v, expected %v", n, length)
+			return nil, fmt.Errorf("insufficient bytes read (-%d, +%d)", n, length)
 		}
 		keys = append(keys, buf)
 	}
 
-	return keys
-}
-
-func postgresRepo(tb testing.TB) (string, string) {
-	postgresImageRef := os.Getenv("CI_POSTGRES_IMAGE")
-	if postgresImageRef == "" {
-		postgresImageRef = defaultPostgresImageRef
-	}
-
-	parts := strings.SplitN(postgresImageRef, ":", 2)
-	if len(parts) != 2 {
-		tb.Fatalf("invalid postgres ref %v", postgresImageRef)
-	}
-	return parts[0], parts[1]
+	return keys, nil
 }