@sethvargo 2020-11-29
Benchmarks on my Mac:

```text
$ go test -count=1 ./pkg/database/...
before: 6.38s user 4.49s system 15% cpu 1:10.91 total
after:  3.03s user 1.94s system 37% cpu 11.408 total
```

```text
$ make test-acc
before: 82.16s user 25.41s system 103% cpu 1:44.14 total
after:  69.93s user 20.64s system 281% cpu 32.165 total
```

Interestingly, the new tests are 3-8x faster, but use ~2.5x the CPU. I'm guessing this is because there's a shared database pool, more tests can run in parallel (using more CPU).