@sethvargo 2020-11-29
Ah, these tests won't pass until we merge and deploy https://github.com/GoogleCloudPlatform/oss-test-infra/pull/626. `WITH (FORCE)` is a postgres 13 thing, but our envvars are still requesting Postgres 12. 

Good news is that cleanup is the _last_ step, meaning it's very likely this shaved a full minute off of CI runtime, but we won't know for certain until that prow config is merged and deployed.

/cc @chaodaiG